refactor: ♻️ (`web`) ランディングページ向けのコンポーネントを`src/features/landingPage`以下に移した。  (#90)
