+++ apps/web/src/app/page.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react';
-import { Hero } from '@/components/Hero/Hero';
 import { JsonLdScript } from '@/components/JsonLdScript/JsonLdScript';
 import { MyPhotoCardStack } from '@/components/MyPhotoCardStack/MyPhotoCardStack';
+import { Hero } from '@/features/landingPage/components/Hero/Hero';
 import { css } from 'styled-system/css';
 
 const Home = (): ReactNode => {
