+++ apps/web/src/features/landingPage/components/Globe/Globe.tsx
