+++ apps/web/src/features/landingPage/components/PhotoCard/PhotoCard.tsx
